Revert "Add integration tests for API endpoints"
