--- .env.test
@@ -1,6 +1,0 @@
-DB_HOST=localhost
-DB_PORT=5432
-DB_USERNAME=postgres
-DB_PASSWORD=postgres
-DB_NAME=test_db
-NODE_ENV=test
