+++ docker-compose.yml
@@ -13,7 +13,6 @@ services:
       - '5432:5432'
     volumes:
       - pgdata:/var/lib/postgresql/data
-      - ./init:/docker-entrypoint-initdb.d
 
   api:
     build: .
