+++ src/config/typeOrm/dataSource.ts
@@ -5,16 +5,13 @@ import { Document } from '../../app/document/entities/Document.entity';
 
 export const AppDataSource = new DataSource({
     type: 'postgres',
-    host: process.env.DB_HOST || 'localhost',
-    port: Number(process.env.DB_PORT) || 5432,
-    username: process.env.DB_USERNAME || 'postgres',
-    password: process.env.DB_PASSWORD || 'postgres',
-    database:
-        process.env.NODE_ENV === 'test'
-            ? process.env.TEST_DB_NAME || 'test_db'
-            : process.env.DB_NAME || 'postgres',
+    host: process.env.DB_HOST,
+    port: 5432,
+    username: process.env.DB_USERNAME,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
     synchronize: true,
     logging: false,
     entities: [DocumentType, Employee, Document],
-    migrations: ['src/config/typeOrm/migrations/*.ts'],
+    migrations: ['src/config/typeOrm/migrations/*.ts']
 });
